Dev/products/Idea_11_Code_Review_Bias_Detector/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
google-generativeai
PyGithub
pandas
numpy
scipy
python-multipart
requests
pytest
stripe
python-dotenv
supabase
redis
reportlab
python-gitlab
